cafeteria/requirements.txt
asgiref==3.4.1
astroid==2.7.3
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.2
platformdirs==2.3.0
pylint==2.10.2
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
